修改用电发电类型的模糊查询

+++ yn-admin/src/main/java/com/yn/web/AAATestController.java
@@ -26,7 +26,7 @@ public class AAATestController {
     public Object getTodayKwh2() {
 
 
-        List<EachHourTemStation> todayKwh2 = temStationService.getTodayKwh(null, AmmeterTypeEnum.GENERATED_ELECTRICITY.getCode());
+        List<EachHourTemStation> todayKwh2 = temStationService.getTodayKwh(null, AmmeterTypeEnum.GENERATED_ELECTRICITY.getCode().longValue());
 
 
         return ResultVOUtil.success(todayKwh2);
